% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fasttau}
\alias{fasttau}
\title{Compute Kendall's Tau-A}
\usage{
fasttau(x, y)
}
\arguments{
\item{x}{A vector of values}

\item{y}{A vector of values (each y-value forms a pair with the corresponding x-value)}
}
\description{
Compute Kendall's Tau-A
}
